Revert "Swipe front"
